requirements.txt
transformers==4.53.0
backoff==2.2.1
openai==1.93.0
chromadb==1.0.15
sentence-transformers==5.0.0
tensorflow==2.18.0
torch==2.5.0
scikit-learn==1.6.1
jax==0.4.28
xgboost==2.1.0
tqdm==4.67.1
python-dotenv==1.1.1